Create 4. Superstore Sales Analysis and Visualization
The goal of this analysis is to generate some insight from the data and answer some business questions. The question can like Top product in terms of sales, Which Region is most profitable etc. Task is to analyze the sales data and identify weak areas and opportunities for Super Store to boost business growth.